fix: #26 player.stop()을 통해 해결 player가 PrepareToPlay상태에서는 playbackTime이 duration과 같아진다.
stop을 해주면 0으로 돌아옴
