+++ Music-Player/Music-Player/View/ContentView.swift
@@ -22,7 +22,6 @@ struct ContentView: View {
                 NavigationView {
                     LibraryView(player: player)
                         .padding(.bottom, 80)
-                        .padding(.top, 20)
                 }
                 .edgesIgnoringSafeArea(.bottom)
                 BlurView()
@@ -62,30 +61,4 @@ struct BlurView: UIViewRepresentable {
     func updateUIView(_ uiView: UIViewType, context: Context) {
         
     }
-}
-
-struct NavigationBarColor: ViewModifier {
-
-  init(backgroundColor: UIColor, tintColor: UIColor) {
-    let coloredAppearance = UINavigationBarAppearance()
-    coloredAppearance.configureWithOpaqueBackground()
-    coloredAppearance.backgroundColor = backgroundColor
-    coloredAppearance.titleTextAttributes = [.foregroundColor: tintColor]
-    coloredAppearance.largeTitleTextAttributes = [.foregroundColor: tintColor]
-                   
-    UINavigationBar.appearance().standardAppearance = coloredAppearance
-    UINavigationBar.appearance().scrollEdgeAppearance = coloredAppearance
-    UINavigationBar.appearance().compactAppearance = coloredAppearance
-    UINavigationBar.appearance().tintColor = tintColor
-  }
-
-  func body(content: Content) -> some View {
-    content
-  }
-}
-
-extension View {
-  func navigationBarColor(backgroundColor: UIColor, tintColor: UIColor) -> some View {
-    self.modifier(NavigationBarColor(backgroundColor: backgroundColor, tintColor: tintColor))
-  }
 }
